requirements.txt
# For a reproducible development environment

# Core Dependencies 
numpy>=1.21
cupy-cuda12x>=12.0

# Benchmark & Dashboard Dependencies
scikit-learn>=1.3
streamlit>=1.30
psutil>=5.9